+++ CHANGES.md
@@ -10,6 +10,8 @@
   'successful-ok-ignored-or-substituted-attributes' status code (Issue #60)
 - Now allow empty "printer-alert" values since PWG 5100.9 is ambiguous about
   how to report the absence of alerts (Issue #61)
+- The "ippevesubmit" tool now supports the "-r" (replay) option for replaying
+  the results of a test (Issue #65)
 - The "dnssd-tests.sh" script did not correctly report errors on monochrome
   printers.